api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain-google-genai==4.1.3
google-generativeai==0.8.6
langchain-core==1.2.7
